build: fix 64-bit compiler warnings in native library
Resolved several C++ compiler warnings related to 64-bit portability
and type casting. Specifically:
- Fixed cast from DWORD to void* in GThreads.cpp using DWORD_PTR.
- Fixed cast from void* to long in debug.cpp using size_t.

+++ libdjvu/debug.cpp
@@ -186,7 +186,7 @@ DjVuDebug::lock(int lvl, int noindent)
   int threads_num=1;
   debug_lock.lock();
   // Get per-thread debug object
-  long threadid = (long) GThread::current();
+  long threadid = (long)(size_t) GThread::current();
   DjVuDebug &dbg = debug_map()[threadid];
   threads_num=debug_map().size();
   // Check level
